pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sia-app"
version = "1.0.0"
description = "Application to facilitate the download, exploration and visual analysis of oceanographic data."
readme = "README.md"
authors = [{ name = "Alexis Montejo, Julian Maldonado et al." }]
maintainers = [
  {name = "SIA Project"}
]
license = { file = "COPYING" }
keywords = [
  "gis",
  "cartography",
  "oceanographic",
  "oceanic",
  "data",
  "visualization",
  "extraction",
  "etl",
  "analysis"
]
dependencies = [
  "ttkbootstrap >= 1.10.1",
  "siaextractlib >= 0.2.2",
  "siaplotlib >= 0.2.1",
  "pillow >= 9.3.0",
  "xarray >= 2023.5.0",
  "numpy >= 1.23.4"
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["bumpver", "build", "twine"]

[project.urls]
Homepage = "https://sia-information-system.github.io/sia-website"
"Source Code" = "https://github.com/sia-information-system/sia-desktop-app"

[project.scripts]
sia-app-debug = "sia_app.__main__:main"

[project.gui-scripts]
sia-app = "sia_app.__main__:main"

[tool.bumpver]
current_version = "1.0.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "Bump version {old_version} -> {new_version}"
commit = true
tag = false
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
  'current_version = "{version}"',
  'version = "{version}"'
]
"README.md" = [
  "**Version:** {version}"
]
"src/sia_app/__init__.py" = [
  "__version__ = '{version}'"
]
